+++ src/main/java/com/example/bnta/bookshelf/controllers/AuthorController.java
@@ -1,0 +1,4 @@
+package com.example.bnta.bookshelf.controllers;
+
+public class AuthorController {
+}
