+++ src/main/java/com/example/bnta/bookshelf/controllers/BookController.java
@@ -1,0 +1,4 @@
+package com.example.bnta.bookshelf.controllers;
+
+public class BookController {
+}
