+++ src/main/java/com/example/bnta/bookshelf/models/BookshelfDTO.java
@@ -1,4 +1,4 @@
-package models;
+package com.example.bnta.bookshelf.models;
 
 public class BookshelfDTO {
