+++ src/main/java/com/example/bnta/bookshelf/models/Genre.java
@@ -1,4 +1,4 @@
-package models;
+package com.example.bnta.bookshelf.models;
 
 public enum Genre { // F & G
         ROMANCE,
