+++ src/main/java/com/example/bnta/bookshelf/repositories/AuthorRepository.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.example.bnta.bookshelf.repositories;
 
-import models.Author;
+import com.example.bnta.bookshelf.models.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuthorRepository extends JpaRepository<Author, Long> {
