+++ src/main/java/com/example/bnta/bookshelf/repositories/BookRepository.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.example.bnta.bookshelf.repositories;
 
-import models.Book;
+import com.example.bnta.bookshelf.models.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository <Book, Long>{
