+++ src/main/java/com/example/bnta/bookshelf/repositories/BookshelfRepository.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.example.bnta.bookshelf.repositories;
 
-import models.Bookshelf;
+import com.example.bnta.bookshelf.models.Bookshelf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookshelfRepository extends JpaRepository <Bookshelf, Long>{
