+++ src/main/java/com/example/bnta/bookshelf/repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.example.bnta.bookshelf.repositories;
 
-import models.User;
+import com.example.bnta.bookshelf.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository <User, Long>{
