+++ src/main/java/com/example/bnta/bookshelf/services/AuthorService.java
@@ -1,0 +1,4 @@
+package com.example.bnta.bookshelf.services;
+
+public class AuthorService {
+}
