+++ src/main/java/com/example/bnta/bookshelf/services/UserService.java
@@ -1,13 +1,12 @@
-package services;
+package com.example.bnta.bookshelf.services;
 
-import models.Bookshelf;
-import models.User;
-import models.UserDTO;
+import com.example.bnta.bookshelf.models.Bookshelf;
+import com.example.bnta.bookshelf.models.User;
+import com.example.bnta.bookshelf.models.UserDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import repositories.BookRepository;
-import repositories.BookshelfRepository;
-import repositories.UserRepository;
+import com.example.bnta.bookshelf.repositories.BookshelfRepository;
+import com.example.bnta.bookshelf.repositories.UserRepository;
 
 import java.util.List;
