--- src/main/java/controllers/AuthorController.java
@@ -1,4 +1,0 @@
-package controllers;
-
-public class AuthorController {
-}
