--- src/main/java/controllers/BookController.java
@@ -1,4 +1,0 @@
-package controllers;
-
-public class BookController {
-}
