--- src/main/java/controllers/GenreController.java
@@ -1,4 +1,0 @@
-package controllers;
-
-public class GenreController {
-}
