--- src/main/java/services/AuthorService.java
@@ -1,4 +1,0 @@
-package services;
-
-public class AuthorService {
-}
