--- src/main/java/services/BookService.java
@@ -1,4 +1,0 @@
-package services;
-
-public class BookService {
-}
