--- src/main/java/services/GenreService.java
@@ -1,4 +1,0 @@
-package services;
-
-public class GenreService {
-}
